% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggamma.R
\name{dggamma}
\alias{dggamma}
\title{Generalized Gamma density (Stacy form with f(t) = k * beta^(k\emph{alpha) / Gamma(alpha) * t^(k}alpha-1) * exp(-(beta*t)^k))}
\usage{
dggamma(x, alpha, beta, kappa, log = FALSE)
}
\arguments{
\item{x}{Numeric vector of quantiles (x > 0).}

\item{alpha}{Shape parameter (>0).}

\item{beta}{Scale-rate parameter (>0) (appears as (beta*x)^kappa).}

\item{kappa}{Family/shape parameter (>0).}

\item{log}{Logical; if TRUE, return log-density.}
}
\value{
Numeric vector of (log-)densities of the same length as x.
}
\description{
Density of the generalized gamma distribution with parameters
alpha (>0), beta (>0), kappa (>0):
f(t) = \link{kappa * beta^(kappa*alpha) / Gamma(alpha)} * t^(kappa\emph{alpha - 1) * exp(-(beta}t)^kappa),  t>0.
}
\examples{
dggamma(1, alpha=2, beta=1, kappa=0.5)
dggamma(c(0.5,1,2), alpha=3, beta=2, kappa=1.5, log=TRUE)

}
